Rename package 'junit_exporter' to 'junit'
