+++ common/junit/include/junit/test_suites.hpp
@@ -16,7 +16,7 @@
 #define JUNIT_EXPORTER__JUNIT_EXPORTER_HPP_
 
 #include <boost/filesystem/path.hpp>
-#include <junit_exporter/test_suite.hpp>
+#include <junit/test_suite.hpp>
 #include <numeric>
 #include <pugixml.hpp>
 #include <string>
